misc renaming & more verbose error handling for missing IDs in metadata
