+++ q2_composition/_ancombc.py
@@ -69,11 +69,19 @@ def _ancombc(table, metadata, formula, p_adj_method, prv_cut, lib_cut,
 
     meta = metadata.to_dataframe()
 
+    # error on IDs found in table but not in metadata
+    missing_ids = []
+    for i in set(table.index):
+        if i not in set(meta.index):
+            missing_ids.append(i)
+
     if not(set(table.index).issubset(set(meta.index))):
         raise KeyError('Not all samples present within the table were found in'
                        ' the associated metadata file. Please make sure that'
                        ' all samples in the FeatureTable are also present in'
-                       ' the metadata.')
+                       ' the metadata.'
+                       ' Sample IDs not found in the metadata: %s'
+                       % missing_ids)
 
     # column validation for the group parameter
     if group is not None:
